+++ simvue/client.py
@@ -107,18 +107,19 @@ class Client:
         server_url : str, optional
             specify URL, if unset this is read from the config file
         """
-        self._config = SimvueConfiguration.fetch(
-            server_token=server_token, server_url=server_url
+        self._user_config = SimvueConfiguration.fetch(
+            server_token=server_token, server_url=server_url, mode="online"
         )
 
         for label, value in zip(
-            ("URL", "API token"), (self._config.server.url, self._config.server.url)
+            ("URL", "API token"),
+            (self._user_config.server.url, self._user_config.server.url),
         ):
             if not value:
                 logger.warning(f"No {label} specified")
 
         self._headers: dict[str, str] = {
-            "Authorization": f"Bearer {self._config.server.token}"
+            "Authorization": f"Bearer {self._user_config.server.token}"
         }
 
     def _get_json_from_response(
@@ -129,6 +130,7 @@ class Client:
     ) -> typing.Union[dict, list]:
         try:
             json_response = response.json()
+            json_response = json_response or {}
         except json.JSONDecodeError:
             json_response = None
 
@@ -183,7 +185,9 @@ class Client:
         params: dict[str, str] = {"filters": json.dumps([f"name == {name}"])}
 
         response: requests.Response = requests.get(
-            f"{self._config.server.url}/api/runs", headers=self._headers, params=params
+            f"{self._user_config.server.url}/api/runs",
+            headers=self._headers,
+            params=params,
         )
 
         json_response = self._get_json_from_response(
@@ -233,7 +237,7 @@ class Client:
         """
 
         response: requests.Response = requests.get(
-            f"{self._config.server.url}/api/runs/{run_id}", headers=self._headers
+            f"{self._user_config.server.url}/api/runs/{run_id}", headers=self._headers
         )
 
         json_response = self._get_json_from_response(
@@ -353,7 +357,9 @@ class Client:
         }
 
         response = requests.get(
-            f"{self._config.server.url}/api/runs", headers=self._headers, params=params
+            f"{self._user_config.server.url}/api/runs",
+            headers=self._headers,
+            params=params,
         )
 
         response.raise_for_status()
@@ -402,7 +408,7 @@ class Client:
         """
 
         response = requests.delete(
-            f"{self._config.server.url}/api/runs/{run_id}",
+            f"{self._user_config.server.url}/api/runs/{run_id}",
             headers=self._headers,
         )
 
@@ -438,7 +444,7 @@ class Client:
         params: dict[str, str] = {"filters": json.dumps([f"path == {path}"])}
 
         response: requests.Response = requests.get(
-            f"{self._config.server.url}/api/folders",
+            f"{self._user_config.server.url}/api/folders",
             headers=self._headers,
             params=params,
         )
@@ -483,7 +489,7 @@ class Client:
         params: dict[str, bool] = {"runs_only": True, "runs": True}
 
         response = requests.delete(
-            f"{self._config.server.url}/api/folders/{folder_id}",
+            f"{self._user_config.server.url}/api/folders/{folder_id}",
             headers=self._headers,
             params=params,
         )
@@ -541,15 +547,14 @@ class Client:
                 return None
             else:
                 raise RuntimeError(
-                    f"Deletion of folder '{folder_path}' failed, "
-                    "folder does not exist."
+                    f"Deletion of folder '{folder_path}' failed, folder does not exist."
                 )
 
         params: dict[str, bool] = {"runs": True} if remove_runs else {}
         params |= {"recursive": recursive}
 
         response = requests.delete(
-            f"{self._config.server.url}/api/folders/{folder_id}",
+            f"{self._user_config.server.url}/api/folders/{folder_id}",
             headers=self._headers,
             params=params,
         )
@@ -580,7 +585,8 @@ class Client:
             the unique identifier for the alert
         """
         response = requests.delete(
-            f"{self._config.server.url}/api/alerts/{alert_id}", headers=self._headers
+            f"{self._user_config.server.url}/api/alerts/{alert_id}",
+            headers=self._headers,
         )
 
         if response.status_code == http.HTTPStatus.OK:
@@ -615,7 +621,7 @@ class Client:
         params: dict[str, str] = {"runs": json.dumps([run_id])}
 
         response: requests.Response = requests.get(
-            f"{self._config.server.url}/api/artifacts",
+            f"{self._user_config.server.url}/api/artifacts",
             headers=self._headers,
             params=params,
         )
@@ -641,7 +647,7 @@ class Client:
         params: dict[str, str | None] = {"name": name}
 
         response = requests.get(
-            f"{self._config.server.url}/api/runs/{run_id}/artifacts",
+            f"{self._user_config.server.url}/api/runs/{run_id}/artifacts",
             headers=self._headers,
             params=params,
         )
@@ -683,7 +689,7 @@ class Client:
         body: dict[str, str | None] = {"id": run_id, "reason": reason}
 
         response = requests.put(
-            f"{self._config.server.url}/api/runs/abort",
+            f"{self._user_config.server.url}/api/runs/abort",
             headers=self._headers,
             json=body,
         )
@@ -877,7 +883,7 @@ class Client:
         params: dict[str, typing.Optional[str]] = {"category": category}
 
         response: requests.Response = requests.get(
-            f"{self._config.server.url}/api/runs/{run_id}/artifacts",
+            f"{self._user_config.server.url}/api/runs/{run_id}/artifacts",
             headers=self._headers,
             params=params,
         )
@@ -903,8 +909,7 @@ class Client:
                     future.result()
                 except Exception as e:
                     raise RuntimeError(
-                        f"Download of file {download['url']} "
-                        f"failed with exception: {e}"
+                        f"Download of file {download['url']} failed with exception: {e}"
                     )
 
     @prettify_pydantic
@@ -973,7 +978,7 @@ class Client:
         }
 
         response: requests.Response = requests.get(
-            f"{self._config.server.url}/api/folders",
+            f"{self._user_config.server.url}/api/folders",
             headers=self._headers,
             params=params,
         )
@@ -1020,7 +1025,7 @@ class Client:
         params = {"runs": json.dumps([run_id])}
 
         response: requests.Response = requests.get(
-            f"{self._config.server.url}/api/metrics/names",
+            f"{self._user_config.server.url}/api/metrics/names",
             headers=self._headers,
             params=params,
         )
@@ -1045,9 +1050,9 @@ class Client:
         run_ids: list[str],
         xaxis: str,
         aggregate: bool,
-        max_points: int = -1,
+        max_points: typing.Optional[int] = None,
     ) -> dict[str, typing.Any]:
-        params: dict[str, typing.Union[str, int]] = {
+        params: dict[str, typing.Union[str, int, None]] = {
             "runs": json.dumps(run_ids),
             "aggregate": aggregate,
             "metrics": json.dumps(metric_names),
@@ -1056,14 +1061,14 @@ class Client:
         }
 
         metrics_response: requests.Response = requests.get(
-            f"{self._config.server.url}/api/metrics",
+            f"{self._user_config.server.url}/api/metrics",
             headers=self._headers,
             params=params,
         )
 
         json_response = self._get_json_from_response(
             expected_status=[http.HTTPStatus.OK],
-            scenario=f"Retrieval of metrics '{metric_names}' in " f"runs '{run_ids}'",
+            scenario=f"Retrieval of metrics '{metric_names}' in runs '{run_ids}'",
             response=metrics_response,
         )
 
@@ -1169,7 +1174,7 @@ class Client:
             run_ids=run_ids,
             xaxis=xaxis,
             aggregate=aggregate,
-            max_points=max_points or -1,
+            max_points=max_points,
         )
 
         if not run_metrics:
@@ -1235,8 +1240,7 @@ class Client:
 
         if data is None:
             raise RuntimeError(
-                f"Cannot plot metrics {metric_names}, "
-                f"no data found for runs {run_ids}."
+                f"Cannot plot metrics {metric_names}, no data found for runs {run_ids}."
             )
 
         # Undo multi-indexing
@@ -1318,7 +1322,7 @@ class Client:
         }
 
         response = requests.get(
-            f"{self._config.server.url}/api/events",
+            f"{self._user_config.server.url}/api/events",
             headers=self._headers,
             params=params,
         )
@@ -1375,7 +1379,7 @@ class Client:
         params: dict[str, int] = {"count": count_limit or 0, "start": start_index or 0}
         if not run_id:
             response = requests.get(
-                f"{self._config.server.url}/api/alerts/",
+                f"{self._user_config.server.url}/api/alerts/",
                 headers=self._headers,
                 params=params,
             )
@@ -1387,7 +1391,7 @@ class Client:
             )
         else:
             response = requests.get(
-                f"{self._config.server.url}/api/runs/{run_id}",
+                f"{self._user_config.server.url}/api/runs/{run_id}",
                 headers=self._headers,
                 params=params,
             )
@@ -1464,7 +1468,9 @@ class Client:
         """
         params = {"count": count_limit or 0, "start": start_index or 0}
         response = requests.get(
-            f"{self._config.server.url}/api/tags", headers=self._headers, params=params
+            f"{self._user_config.server.url}/api/tags",
+            headers=self._headers,
+            params=params,
         )
 
         json_response = self._get_json_from_response(
@@ -1498,7 +1504,7 @@ class Client:
         """
 
         response = requests.delete(
-            f"{self._config.server.url}/api/tags/{tag_id}",
+            f"{self._user_config.server.url}/api/tags/{tag_id}",
             headers=self._headers,
         )
 
@@ -1540,7 +1546,7 @@ class Client:
         """
 
         response: requests.Response = requests.get(
-            f"{self._config.server.url}/api/tag/{tag_id}", headers=self._headers
+            f"{self._user_config.server.url}/api/tag/{tag_id}", headers=self._headers
         )
 
         json_response = self._get_json_from_response(