+++ tests/functional/test_run_class.py
@@ -1,3 +1,5 @@
+import os
+from os.path import basename
 import pytest
 import pytest_mock
 import time
@@ -15,6 +17,7 @@ import random
 import simvue
 import simvue.run as sv_run
 import simvue.client as sv_cl
+import simvue.sender as sv_send
 
 if typing.TYPE_CHECKING:
     from .conftest import CountingLogHandler
@@ -126,10 +129,28 @@ def test_log_metrics(
 
 
 @pytest.mark.run
-def test_log_metrics_offline(create_test_run_offline: tuple[sv_run.Run, dict]) -> None:
+def test_log_metrics_offline(create_plain_run_offline: tuple[sv_run.Run, dict]) -> None:
     METRICS = {"a": 10, "b": 1.2, "c": 2}
-    run, _ = create_test_run_offline
+    run, _ = create_plain_run_offline
     run.log_metrics(METRICS)
+    run_id, *_ = sv_send.sender()
+    time.sleep(1.0)
+    run.close()
+    client = sv_cl.Client()
+    _data = client.get_metric_values(
+        run_ids=[run_id],
+        metric_names=list(METRICS.keys()),
+        xaxis="step",
+        aggregate=False,
+    )
+    assert sorted(set(METRICS.keys())) == sorted(set(_data.keys()))
+    _steps = []
+    for entry in _data.values():
+        _steps += list(i[0] for i in entry.keys())
+    _steps = set(_steps)
+    assert (
+        len(_steps) == 1
+    )
 
 
 @pytest.mark.run
@@ -137,29 +158,81 @@ def test_log_events(create_test_run: tuple[sv_run.Run, dict]) -> None:
     EVENT_MSG = "Hello world!"
     run, _ = create_test_run
     run.log_event(EVENT_MSG)
+    time.sleep(1.0)
+    run.close()
+    client = sv_cl.Client()
+    event_data = client.get_events(run.id, count_limit=1)
+    assert event_data[0].get("message", EVENT_MSG)
+
 
 
 @pytest.mark.run
-def test_log_events_offline(create_test_run_offline: tuple[sv_run.Run, dict]) -> None:
-    EVENT_MSG = "Hello world!"
-    run, _ = create_test_run_offline
+def test_log_events_offline(create_plain_run_offline: tuple[sv_run.Run, dict]) -> None:
+    EVENT_MSG = "Hello offline world!"
+    run, _ = create_plain_run_offline
     run.log_event(EVENT_MSG)
+    run_id, *_ = sv_send.sender()
+    run.close()
+    time.sleep(1.0)
+    client = sv_cl.Client()
+    event_data = client.get_events(run_id, count_limit=1)
+    assert event_data[0].get("message", EVENT_MSG)
 
 
 @pytest.mark.run
-def test_update_metadata(create_test_run: tuple[sv_run.Run, dict]) -> None:
+def test_offline_tags(create_plain_run_offline: tuple[sv_run.Run, dict]) -> None:
+    run, run_data = create_plain_run_offline
+    run_id, *_ = sv_send.sender()
+    run.close()
+    time.sleep(1.0)
+    client = sv_cl.Client()
+    tags = client.get_tags()
+    assert run_data["tags"][-1] in [tag["name"] for tag in tags]
+
+
+
+@pytest.mark.run
+def test_update_metadata_running(create_test_run: tuple[sv_run.Run, dict]) -> None:
     METADATA = {"a": 10, "b": 1.2, "c": "word"}
     run, _ = create_test_run
     run.update_metadata(METADATA)
+    run.close()
+    time.sleep(1.0)
+    client = sv_cl.Client()
+    run_info = client.get_run(run.id)
+
+    for key, value in METADATA.items():
+        assert run_info.get("metadata", {}).get(key) == value
+
+
+@pytest.mark.run
+def test_update_metadata_created(create_pending_run: tuple[sv_run.Run, dict]) -> None:
+    METADATA = {"a": 10, "b": 1.2, "c": "word"}
+    run, _ = create_pending_run
+    run.update_metadata(METADATA)
+    time.sleep(1.0)
+    client = sv_cl.Client()
+    run_info = client.get_run(run.id)
+
+    for key, value in METADATA.items():
+        assert run_info.get("metadata", {}).get(key) == value
 
 
 @pytest.mark.run
 def test_update_metadata_offline(
-    create_test_run_offline: tuple[sv_run.Run, dict],
+    create_plain_run_offline: tuple[sv_run.Run, dict],
 ) -> None:
     METADATA = {"a": 10, "b": 1.2, "c": "word"}
-    run, _ = create_test_run_offline
+    run, _ = create_plain_run_offline
     run.update_metadata(METADATA)
+    run_id, *_ = sv_send.sender()
+    run.close()
+    time.sleep(1.0)
+    client = sv_cl.Client()
+    run_info = client.get_run(run_id)
+
+    for key, value in METADATA.items():
+        assert run_info.get("metadata", {}).get(key) == value
 
 
 @pytest.mark.run
@@ -364,7 +437,7 @@ def test_set_folder_details(request: pytest.FixtureRequest) -> None:
         run.set_folder_details(path=folder_name, tags=tags, description=description)
 
     client = sv_cl.Client()
-    assert (folder := client.get_folders(filters=[f"path == {folder_name}"])[0])["tags"] == tags
+    assert sorted((folder := client.get_folders(filters=[f"path == {folder_name}"])[0])["tags"]) == sorted(tags)
     assert folder["description"] == description
 
 
@@ -378,13 +451,15 @@ def test_set_folder_details(request: pytest.FixtureRequest) -> None:
 @pytest.mark.parametrize("name", ("test_file", None), ids=("named", "nameless"))
 @pytest.mark.parametrize("allow_pickle", (True, False), ids=("pickled", "unpickled"))
 @pytest.mark.parametrize("empty_file", (True, False), ids=("empty", "content"))
-def test_save_file(
+@pytest.mark.parametrize("category", ("input", "output", "code"))
+def test_save_file_online(
     create_plain_run: typing.Tuple[sv_run.Run, dict],
     valid_mimetype: bool,
     preserve_path: bool,
     name: typing.Optional[str],
     allow_pickle: bool,
     empty_file: bool,
+    category: typing.Literal["input", "output", "code"],
     capfd,
 ) -> None:
     simvue_run, _ = create_plain_run
@@ -399,7 +474,7 @@ def test_save_file(
         if valid_mimetype:
             simvue_run.save_file(
                 out_name,
-                category="input",
+                category=category,
                 filetype=file_type,
                 preserve_path=preserve_path,
                 name=name,
@@ -408,7 +483,7 @@ def test_save_file(
             with pytest.raises(RuntimeError):
                 simvue_run.save_file(
                     out_name,
-                    category="input",
+                    category=category,
                     filetype=file_type,
                     preserve_path=preserve_path,
                 )
@@ -421,10 +496,70 @@ def test_save_file(
                     variable.out
                     == "[simvue] WARNING: saving zero-sized files not currently supported\n"
                 )
+                return
+        simvue_run.close()
+        time.sleep(1.0)
+        os.remove(out_name)
+        client = sv_cl.Client()
+        base_name = name or out_name.name
+        if preserve_path:
+            out_loc = pathlib.Path(tempd) / out_name.parent
+            stored_name = out_name.parent / pathlib.Path(base_name)
+        else:
+            out_loc = pathlib.Path(tempd)
+            stored_name = pathlib.Path(base_name)
+        out_file = out_loc.joinpath(name or out_name.name)
+        client.get_artifact_as_file(run_id=simvue_run.id, name=f"{name or stored_name}", path=tempd)
+        assert out_loc.joinpath(name if name else out_name.name).exists()
 
 
 @pytest.mark.run
-def test_update_tags(
+@pytest.mark.parametrize(
+    "preserve_path", (True, False), ids=("preserve_path", "modified_path")
+)
+@pytest.mark.parametrize("name", ("retrieved_test_file", None), ids=("named", "nameless"))
+@pytest.mark.parametrize("category", ("input", "output", "code"))
+def test_save_file_offline(
+    create_plain_run_offline: tuple[sv_run.Run, dict],
+    preserve_path: bool,
+    name: typing.Optional[str],
+    category: typing.Literal["input", "output", "code"]
+) -> None:
+    simvue_run, _ = create_plain_run_offline
+    file_type: str = "text/plain"
+    with tempfile.TemporaryDirectory() as tempd:
+        with open(
+            (out_name := pathlib.Path(tempd).joinpath("test_file.txt")),
+            "w",
+        ) as out_f:
+            out_f.write("test data entry")
+
+        simvue_run.save_file(
+            out_name,
+            category=category,
+            preserve_path=preserve_path,
+            name=name,
+        )
+        run_id, *_ = sv_send.sender()
+        simvue_run.close()
+        time.sleep(1.0)
+        os.remove(out_name)
+        client = sv_cl.Client()
+        assert run_id
+        base_name = name or out_name.name
+        if preserve_path:
+            out_loc = pathlib.Path(tempd) / out_name.parent
+            stored_name = out_name.parent / pathlib.Path(base_name)
+        else:
+            out_loc = pathlib.Path(tempd)
+            stored_name = pathlib.Path(base_name)
+        out_file = out_loc.joinpath(name or out_name.name)
+        client.get_artifact_as_file(run_id=run_id, name=f"{name or stored_name}", path=tempd)
+        assert out_loc.joinpath(name if name else out_name.name).exists()
+
+
+@pytest.mark.run
+def test_update_tags_running(
     create_plain_run: typing.Tuple[sv_run.Run, dict],
     request: pytest.FixtureRequest,
 ) -> None:
@@ -441,6 +576,32 @@ def test_update_tags(
     client = sv_cl.Client()
     run_data = client.get_run(simvue_run._id)
     assert run_data["tags"] == tags
+
+    simvue_run.update_tags(["additional"])
+
+    time.sleep(1)
+    run_data = client.get_run(simvue_run._id)
+    assert sorted(run_data["tags"]) == sorted(tags + ["additional"])
+
+
+@pytest.mark.run
+def test_update_tags_created(
+    create_pending_run: typing.Tuple[sv_run.Run, dict],
+    request: pytest.FixtureRequest,
+) -> None:
+    simvue_run, _ = create_pending_run
+
+    tags = [
+        "simvue_client_unit_tests",
+        request.node.name.replace("[", "_").replace("]", "_"),
+    ]
+
+    simvue_run.set_tags(tags)
+
+    time.sleep(1)
+    client = sv_cl.Client()
+    run_data = client.get_run(simvue_run._id)
+    assert sorted(run_data["tags"]) == sorted(tags)
 
     simvue_run.update_tags(["additional"])
 
@@ -550,8 +711,11 @@ def test_abort_on_alert_raise(create_plain_run: typing.Tuple[sv_run.Run, dict], 
     run.add_process(identifier="forever_long", executable="bash", c="sleep 10")
     time.sleep(2)
     run.log_alert(alert_id, "critical")
-    time.sleep(4)
-    if not run._status == "terminated":
+    counter = 0
+    while run._status != "terminated" and counter < 15:
+        time.sleep(1)
+        counter += 1
+    if counter >= 15:
         run.kill_all_processes()
         raise AssertionError("Run was not terminated")