+++ CHANGELOG.md
@@ -2,7 +2,24 @@
 
 ## Unreleased
 
+* Add support for defining Simvue run defaults using `tool.simvue` in a project `pyproject.toml` file.
 * Drop support for INI based configuration files.
+* Retrieve all metric values if `max_points` is unspecified or set to `None`.
+## [v1.1.4](https://github.com/simvue-io/client/releases/tag/v1.1.4) - 2024-12-11
+
+* Remove incorrect identifier reference for latest Simvue servers during reconnection.
+* Fixed missing online mode selection when retrieving configuration for `Client` class.
+
+## [v1.1.3](https://github.com/simvue-io/client/releases/tag/v1.1.3) - 2024-12-09
+
+* Fixed bug with `requirements.txt` metadata read.
+* Added Simvue server version check.
+* Remove checking of server version in offline mode and add default run mode to configuration options.
+* Fix offline mode class initialisation, and propagation of configuration.
+
+## [v1.1.2](https://github.com/simvue-io/client/releases/tag/v1.1.2) - 2024-11-06
+
+* Fix bug in offline mode directory retrieval.
 
 ## [v1.1.1](https://github.com/simvue-io/client/releases/tag/v1.1.1) - 2024-10-22